// app/src/generated/models/Csa_tasksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Csa_taskscsa_responsible = {
  100000000: 'Me',
  100000001: 'Customer'
} as const;
export type Csa_taskscsa_responsible = keyof typeof Csa_taskscsa_responsible;
export const Csa_taskscsa_status = {
  100000000: 'Backlog',
  100000001: 'ToDo',
  100000002: 'InProgress',
  100000003: 'Waiting',
  100000004: 'Done'
} as const;
export type Csa_taskscsa_status = keyof typeof Csa_taskscsa_status;
export const Csa_tasksstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Csa_tasksstatecode = keyof typeof Csa_tasksstatecode;
export const Csa_tasksstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Csa_tasksstatuscode = keyof typeof Csa_tasksstatuscode;

export interface Csa_tasksBase {
  csa_description?: string;
  csa_duedate?: string;
  csa_name?: string;
  "csa_ProjectId@odata.bind": string;
  csa_responsible?: Csa_taskscsa_responsible;
  csa_sortorder?: number;
  csa_status?: Csa_taskscsa_status;
  csa_taskid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Csa_tasksstatecode;
  statuscode?: Csa_tasksstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Csa_tasks extends Csa_tasksBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  csa_projectidname?: string;
  csa_responsiblename?: string;
  csa_statusname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  csa_projectid?: object;
  _csa_projectid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}